+++ Sources/KeyboardKit/Keyboard/KeyboardContext.swift
@@ -23,7 +23,6 @@ public protocol KeyboardContext: AnyObject {
     
     var actionHandler: KeyboardActionHandler { get set }
     var keyboardAppearanceProvider: KeyboardAppearanceProvider { get set }
-    var keyboardBehavior: KeyboardBehavior { get set }
     var keyboardInputSetProvider: KeyboardInputSetProvider { get set }
     var keyboardLayoutProvider: KeyboardLayoutProvider { get set }